taking off - getting started with the tutorial
This commit is mainly the Converter and Rocketry
code. The book is a bit hard to follow, at the
moment, I'm just punching in code and sorting it
out as I go.

Put some kala chana from 2008 in the pot to soak.
It will have to soak overnight at the least, maybe
longer. Will then cook it on low heat for many hours.
Will be interesting to see how it turns out.