recursion: done for now
This has been a frustrating chapter. I will allow
that most if not all of my confusion stems from my
unwillingness to puzzle through the story elements
which I found unclear. In a sense, I also have to
admit that it seems similar to a recent experience
where "you're just supposed to know." That is,
smart, experience developers understand the nuance
and implications of how certain specifications are
worded.

My unwillingness is mostly related to the fact that
I'm learning this on my own time, and I have a lot
of other things competing for my attention.