if-unless: pattern matching for guard clause
I really like what Elixir can do as shown by the
construction of the guard clause in this commit.

This was a pretty short chapter. Once again, I'm not
that super on what the author wants, as I haven't taken
much time to puzzle through the meaning.

To be fair to myself, the author isn't providing me
with exactly what I'm looking for either. Hence, I'm
doing the work which corresponds as closely as
possible to what I'm trying to get out of it. What I'm
trying to get out of it is an overview of the syntax
and structure of Elixir and Elixir applications. I do
not want to be sidetracked into application details
at this very moment. I'll probably rework all the tests
after I make the first pass through the material.